rename models to get_models
